requirements.txt
asgiref==3.11.0
boto3==1.42.49
botocore==1.42.49
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
dj-database-url==3.1.0
Django==6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
gunicorn==25.0.3
idna==3.11
jmespath==1.1.0
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
requests==2.32.5
s3transfer==0.16.0
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.5
stripe==14.3.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
whitenoise==6.11.0
